refactor: 모달 폴더 이동
