+++ frontend/src/components/MainPage/Menu/Menu.tsx
@@ -2,7 +2,7 @@ import { FC, useContext, useEffect, useState } from 'react'
 import styled from 'styled-components'
 
 import { Image } from 'src/components/common/Image/Image'
-import OptionSelectModal from 'src/components/common/Modal/OptionSelectModal'
+import OptionSelectModal from 'src/components/common/Modal/OptionSelectModal/OptionSelectModal'
 import { InternationalizationContext } from 'src/contexts/InternationalizationContext'
 import useModal from 'src/hooks/useModal'
 import { ProductOptionType } from 'src/types/api/product'
